Additional JS methods implemented
