+++ gradle/libs.versions.toml
@@ -9,6 +9,8 @@ kotlin = "2.0.0-RC3"
 vanniktech = "0.22.0"
 ktlint = "12.1.0"
 dokka = "1.9.20"
+kotlinx_datetime = "0.6.0-RC.2"
+kotlinx_coroutines = "1.8.1"
 
 [libraries]
 retrofit2 = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit2" }
@@ -37,6 +39,8 @@ mockk = { module = "io.mockk:mockk", version = "1.11.0" }
 owner = { module = "org.aeonbits.owner:owner", version = "1.0.12" }
 mockito = { module = "org.mockito:mockito-core", version = "4.8.1" }
 hamcrest = { module = "org.hamcrest:hamcrest-all", version = "1.3" }
+datetime = { module = "org.jetbrains.kotlinx:kotlinx-datetime", version.ref = "kotlinx_datetime"}
+coroutines-test = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-test", version.ref = "kotlinx_coroutines"}
 
 # plugins for included build
 kotlin-gradlePlugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
