+++ pubnub-core/pubnub-core-api/src/commonMain/kotlin/com/pubnub/api/JsonValue.kt
@@ -1,9 +1,9 @@
 package com.pubnub.api
 
-expect abstract class JsonElement
+expect abstract class JsonValue
 //expect class JsonObject
 
-expect fun JsonElement.asString(): String
+expect fun JsonValue.asString(): String
 //expect fun JsonElement.asNumber(): Number
 //expect fun JsonElement.asMap(): Map<String, Any?>
 //expect fun JsonElement.asList(): List<Any?>
