+++ pubnub-core/pubnub-core-api/src/commonMain/kotlin/com/pubnub/api/models/consumer/presence/PNGetStateResult.kt
@@ -1,6 +1,6 @@
 package com.pubnub.api.models.consumer.presence
 
-import com.pubnub.api.JsonElement
+import com.pubnub.api.JsonValue
 
 /**
  * Result of the GetPresenceState operation.
@@ -8,5 +8,5 @@ import com.pubnub.api.JsonElement
  * @property stateByUUID Map of UUIDs and the user states.
  */
 class PNGetStateResult(
-    val stateByUUID: Map<String, JsonElement>,
+    val stateByUUID: Map<String, JsonValue>,
 )
