+++ pubnub-core/pubnub-core-api/src/commonMain/kotlin/com/pubnub/api/models/consumer/pubsub/MessageResult.kt
@@ -1,10 +1,10 @@
 package com.pubnub.api.models.consumer.pubsub
 
-import com.pubnub.api.JsonElement
+import com.pubnub.api.JsonValue
 
 /**
  * @property message The actual message content
  */
 interface MessageResult : PubSubResult {
-    val message: JsonElement
+    val message: JsonValue
 }
