+++ pubnub-core/pubnub-core-api/src/commonMain/kotlin/com/pubnub/api/models/consumer/pubsub/PNSignalResult.kt
@@ -1,11 +1,11 @@
 package com.pubnub.api.models.consumer.pubsub
 
-import com.pubnub.api.JsonElement
+import com.pubnub.api.JsonValue
 
 /**
  * Wrapper around a received signal.
  */
 data class PNSignalResult(
     private val basePubSubResult: PubSubResult,
-    override val message: JsonElement,
+    override val message: JsonValue,
 ) : MessageResult, PubSubResult by basePubSubResult
