+++ pubnub-core/pubnub-core-api/src/commonMain/kotlin/com/pubnub/api/models/consumer/pubsub/objects/ObjectPayload.kt
@@ -1,11 +1,11 @@
 package com.pubnub.api.models.consumer.pubsub.objects
 
-import com.pubnub.api.JsonElement
+import com.pubnub.api.JsonValue
 
 data class ObjectPayload(
     val source: String,
     val version: String,
     val event: String,
     val type: String,
-    val data: JsonElement,
+    val data: JsonValue,
 )
