+++ pubnub-core/pubnub-core-api/src/jvmMain/kotlin/com/pubnub/api/JsonElement.jvm.kt
@@ -1,7 +1,7 @@
 package com.pubnub.api
 
-actual typealias JsonElement = com.google.gson.JsonElement
+actual typealias JsonValue = com.google.gson.JsonElement
 
-actual fun JsonElement.asString(): String {
+actual fun JsonValue.asString(): String {
     return this.asString
 }
