+++ pubnub-kotlin/pubnub-kotlin-api/build.gradle.kts
@@ -90,12 +90,6 @@ kotlin {
             }
         }
 
-        val commonTest by getting {
-            dependencies {
-                implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.6.0-RC.2")
-            }
-        }
-
         val jvmMain by getting {
             dependencies {
                 implementation(project(":pubnub-core:pubnub-core-impl"))
