+++ pubnub-kotlin/pubnub-kotlin-api/src/commonMain/kotlin/com/pubnub/api/models/consumer/objects/membership/PNChannelMembership.kt
@@ -1,17 +1,18 @@
 package com.pubnub.api.models.consumer.objects.membership
 
+import com.pubnub.kmp.CustomObject
 import com.pubnub.api.models.consumer.objects.channel.PNChannelMetadata
 
 data class PNChannelMembership(
     val channel: PNChannelMetadata?,
-    val custom: Any?,
+    val custom: Map<String,Any?>?,
     val updated: String,
     val eTag: String,
     val status: String?,
 ) {
     data class Partial(
         val channelId: String,
-        override val custom: Any? = null,
+        override val custom: CustomObject? = null,
         override val status: String? = null,
     ) : ChannelMembershipInput {
         override val channel: String = channelId
