+++ pubnub-kotlin/pubnub-kotlin-api/src/commonMain/kotlin/com/pubnub/kmp/abstractions.kt
@@ -1,0 +1,3 @@
+package com.pubnub.kmp
+
+expect class CustomObject
