+++ pubnub-kotlin/pubnub-kotlin-api/src/iosMain/kotlin/com/pubnub/kmp/abstractions.ios.kt
@@ -1,0 +1,3 @@
+package com.pubnub.kmp
+
+actual class CustomObject
