+++ pubnub-kotlin/pubnub-kotlin-api/src/iosMain/kotlin/com/pubnub/kmp/factories.ios.kt
@@ -1,0 +1,41 @@
+package com.pubnub.kmp
+
+import com.pubnub.api.PubNub
+import com.pubnub.api.PubNubImpl
+import com.pubnub.api.models.consumer.PNStatus
+import com.pubnub.api.models.consumer.pubsub.PNMessageResult
+import com.pubnub.api.models.consumer.pubsub.PNPresenceEventResult
+import com.pubnub.api.models.consumer.pubsub.PNSignalResult
+import com.pubnub.api.models.consumer.pubsub.files.PNFileEventResult
+import com.pubnub.api.models.consumer.pubsub.message_actions.PNMessageActionResult
+import com.pubnub.api.models.consumer.pubsub.objects.PNObjectEventResult
+import com.pubnub.api.v2.PNConfiguration
+import com.pubnub.api.v2.callbacks.EventListener
+import com.pubnub.api.v2.callbacks.StatusListener
+
+actual fun createPubNub(config: PNConfiguration): PubNub {
+    return PubNubImpl(config)
+}
+
+actual fun createEventListener(
+    pubnub: PubNub,
+    onMessage: (PubNub, PNMessageResult) -> Unit,
+    onPresence: (PubNub, PNPresenceEventResult) -> Unit,
+    onSignal: (PubNub, PNSignalResult) -> Unit,
+    onMessageAction: (PubNub, PNMessageActionResult) -> Unit,
+    onObjects: (PubNub, PNObjectEventResult) -> Unit,
+    onFile: (PubNub, PNFileEventResult) -> Unit
+): EventListener {
+    TODO("Not yet implemented")
+}
+
+actual fun createStatusListener(
+    pubnub: PubNub,
+    onStatus: (PubNub, PNStatus) -> Unit
+): StatusListener {
+    TODO("Not yet implemented")
+}
+
+actual fun createCustomObject(map: Map<String, Any?>): CustomObject {
+    TODO("Not yet implemented")
+}
