+++ pubnub-kotlin/pubnub-kotlin-api/src/jsMain/kotlin/com/pubnub/api/PubNub.js.kt
@@ -44,13 +44,11 @@ import com.pubnub.api.endpoints.push.RemoveChannelsFromPush
 import com.pubnub.api.enums.PNPushEnvironment
 import com.pubnub.api.enums.PNPushType
 import com.pubnub.api.models.consumer.PNBoundedPage
-import com.pubnub.api.models.consumer.PNStatus
 import com.pubnub.api.models.consumer.access_manager.sum.SpacePermissions
 import com.pubnub.api.models.consumer.access_manager.sum.UserPermissions
 import com.pubnub.api.models.consumer.access_manager.v3.ChannelGrant
 import com.pubnub.api.models.consumer.access_manager.v3.ChannelGroupGrant
 import com.pubnub.api.models.consumer.access_manager.v3.UUIDGrant
-import com.pubnub.api.models.consumer.files.PNDownloadableFile
 import com.pubnub.api.models.consumer.message_actions.PNMessageAction
 import com.pubnub.api.models.consumer.objects.PNKey
 import com.pubnub.api.models.consumer.objects.PNMemberKey
@@ -61,21 +59,10 @@ import com.pubnub.api.models.consumer.objects.member.MemberInput
 import com.pubnub.api.models.consumer.objects.member.PNUUIDDetailsLevel
 import com.pubnub.api.models.consumer.objects.membership.ChannelMembershipInput
 import com.pubnub.api.models.consumer.objects.membership.PNChannelDetailsLevel
-import com.pubnub.api.models.consumer.pubsub.BasePubSubResult
-import com.pubnub.api.models.consumer.pubsub.PNMessageResult
-import com.pubnub.api.models.consumer.pubsub.PNPresenceEventResult
-import com.pubnub.api.models.consumer.pubsub.PNSignalResult
-import com.pubnub.api.models.consumer.pubsub.files.PNFileEventResult
-import com.pubnub.api.models.consumer.pubsub.message_actions.PNMessageActionResult
-import com.pubnub.api.models.consumer.pubsub.objects.PNObjectEventResult
 import com.pubnub.api.v2.PNConfiguration
 import com.pubnub.api.v2.callbacks.EventListener
 import com.pubnub.api.v2.callbacks.StatusListener
-import PubNub as PubNubJs
-
-actual fun createPubNub(config: PNConfiguration): PubNub {
-    return PubNubImpl(config)
-}
+import com.pubnub.kmp.CustomObject
 
 actual interface PubNub {
 
@@ -140,7 +127,7 @@ actual interface PubNub {
 
     actual fun whereNow(uuid: String): WhereNow
     actual fun setPresenceState(
-        channels: List<String>, channelGroups: List<String>, state: Any, uuid: String
+        channels: List<String>, channelGroups: List<String>, state: Any
     ): SetState
 
     actual fun getPresenceState(
@@ -368,112 +355,4 @@ actual interface PubNub {
     actual fun setToken(token: String?)
     actual fun destroy()
 
-}
-
-actual fun createEventListener(
-    pubnub: PubNub,
-    onMessage: (PubNub, PNMessageResult) -> Unit,
-    onPresence: (PubNub, PNPresenceEventResult) -> Unit,
-    onSignal: (PubNub, PNSignalResult) -> Unit,
-    onMessageAction: (PubNub, PNMessageActionResult) -> Unit,
-    onObjects: (PubNub, PNObjectEventResult) -> Unit,
-    onFile: (PubNub, PNFileEventResult) -> Unit
-): EventListener {
-    return object : PubNubJs.ListenerParameters {
-        override val message: (PubNubJs.MessageEvent) -> Unit = { messageEvent ->
-            onMessage(pubnub, PNMessageResult(
-                    BasePubSubResult(
-                        messageEvent.channel,
-                        messageEvent.subscription,
-                        messageEvent.timetoken.toLong(),
-                        messageEvent.userMetadata as? JsonElement, // TODO kmp
-                        messageEvent.publisher
-                    ),
-                    messageEvent.message.unsafeCast<JsonElement>(), // TODO kmp
-                    null //TODO kmp error
-                ))
-        }
-        override val presence: ((presenceEvent: PubNubJs.PresenceEvent) -> Unit) = { presenceEvent ->
-            onPresence(pubnub, PNPresenceEventResult(
-                presenceEvent.action,
-                presenceEvent.uuid,
-                presenceEvent.timestamp.toLong(),
-                presenceEvent.occupancy.toInt(),
-                presenceEvent.state.asDynamic(), //TODO kmp
-                presenceEvent.channel,
-                presenceEvent.subscription,
-                presenceEvent.timetoken.toLong(),
-            ))
-
-        }
-        override val signal: ((signalEvent: PubNubJs.SignalEvent) -> Unit) = { signalEvent ->
-            onSignal(pubnub, PNSignalResult(
-                BasePubSubResult(
-                    signalEvent.channel,
-                    signalEvent.subscription,
-                    signalEvent.timetoken.toLong(),
-                    null,
-                    signalEvent.publisher
-                ),
-                signalEvent.message as JsonElement, // TODO kmp
-            ))
-        }
-
-        override val messageAction: ((messageActionEvent: PubNubJs.MessageActionEvent) -> Unit) = { messageActionEvent ->
-            onMessageAction(pubnub, PNMessageActionResult(
-                BasePubSubResult(
-                    messageActionEvent.channel,
-                    messageActionEvent.subscription,
-                    messageActionEvent.timetoken.toLong(),
-                    null,
-                    messageActionEvent.publisher
-                ),
-                messageActionEvent.event,
-                PNMessageAction(
-                    messageActionEvent.data.type,
-                    messageActionEvent.data.value,
-                    messageActionEvent.data.messageTimetoken.toLong()
-                ).apply {
-                    actionTimetoken = messageActionEvent.data.messageTimetoken.toLong()
-                }
-            ))
-        }
-        override val file: ((fileEvent: PubNubJs.FileEvent) -> Unit) = { fileEvent ->
-            onFile(pubnub, PNFileEventResult(
-                fileEvent.channel,
-                fileEvent.timetoken.toLong(),
-                fileEvent.publisher,
-                fileEvent.message,
-                PNDownloadableFile(fileEvent.file.id, fileEvent.file.name, fileEvent.file.url),
-                fileEvent.message as JsonElement, //TODO kmp
-                fileEvent.subscription,
-                null // TODO kmp error
-            ) )
-        }
-    }
-}
-
-actual fun createStatusListener(
-    pubnub: PubNub,
-    onStatus: (PubNub, PNStatus) -> Unit
-): StatusListener {
-    return object : PubNubJs.StatusListenerParameters {
-        override val status: ((statusEvent: PubNubJs.StatusEvent) -> Unit) = { statusEvent ->
-            onStatus(pubnub, PNStatus(
-                enumValueOf(statusEvent.category), //TODO parse category
-                null,
-                statusEvent.currentTimetoken.toString().toLong(),
-                statusEvent.affectedChannels.toList(),
-                statusEvent.affectedChannelGroups.toList()
-            ))
-        }
-    }
-}
-
-class CustomObjectImpl(map: Map<String,Any?> = emptyMap()) : Map<String, Any?> by map
-
-actual typealias CustomObject = CustomObjectImpl
-
-actual fun createCustomObject(map: Map<String, Any?>): CustomObject {
-    return CustomObjectImpl(map)
 }
