+++ pubnub-kotlin/pubnub-kotlin-api/src/jsMain/kotlin/com/pubnub/api/endpoints/presence/HereNowImpl.kt
@@ -2,7 +2,7 @@ package com.pubnub.api.endpoints.presence
 
 import PubNub
 import com.pubnub.api.EndpointImpl
-import com.pubnub.api.JsonElement
+import com.pubnub.api.JsonValue
 import com.pubnub.api.models.consumer.presence.PNHereNowChannelData
 import com.pubnub.api.models.consumer.presence.PNHereNowOccupantData
 import com.pubnub.api.models.consumer.presence.PNHereNowResult
@@ -22,7 +22,7 @@ class HereNowImpl(pubnub: PubNub, params: PubNub.HereNowParameters) : HereNow, E
                     data.occupants.map { occupant ->
                         PNHereNowOccupantData(
                             occupant.uuid,
-                            occupant.state as JsonElement // TODO kmp
+                            occupant.state as JsonValue // TODO kmp
                         )
                     }
                 )
