+++ pubnub-kotlin/pubnub-kotlin-api/src/jsMain/kotlin/com/pubnub/kmp/abstractions.js.kt
@@ -1,0 +1,5 @@
+package com.pubnub.kmp
+
+actual typealias CustomObject = CustomObjectImpl
+
+class CustomObjectImpl(map: Map<String,Any?> = emptyMap()) : Map<String, Any?> by map
