+++ pubnub-kotlin/pubnub-kotlin-api/src/jvmMain/kotlin/com/pubnub/kmp/abstractions.jvm.kt
@@ -1,0 +1,3 @@
+package com.pubnub.kmp
+
+actual typealias CustomObject = Any
